Dilemma of poison
The Prime Minister of a country invites 1000 senators to his annual party. As a tradition, each senator brings the PM a bottle of wine. Soon after, the CBI discovers that one of the senators is trying to assassinate the PM by giving him a bottle of poisoned wine. Unfortunately, they do not know which senator, nor which bottle of wine is poisoned, and the poison is completely indiscernible. However, the PM has some testers he plans to taste wine. He decides to use them as taste testers to determine which bottle of wine contains the poison. The poison when taken has no effect on the prisoner until exactly 24 hours later when the infected tester suddenly dies. The PM needs to determine which bottle of wine is poisoned by tomorrow so that the festivities can continue as planned. Hence he only has time for one round of testing. How many least testers the PM can use to ensure that 24 hours from now he is guaranteed to have found the poisoned wine bottle?